fix warning C4777: 'fprintf' : format string '%lu' requires an argument of type 'unsigned long', but variadic argument 3 has type 'slist_index'; consider using '%u' in the format string; consider using '%I32u' in the format string

+++ src/test/slist_test.c
@@ -47,7 +47,7 @@ static int
 _check_len( slist *a, slist_index expected, const char *fn, int line )
 {
 	if ( a->n == expected ) return 1;
-	fprintf( stderr, "Failed: %s() line %d: Expected slist length of %lu, found %lu\n", fn, line, expected, a->n );
+	fprintf( stderr, "Failed: %s() line %d: Expected slist length of %u, found %u\n", fn, line, (unsigned int)expected, (unsigned int)a->n );
 	return 0;
 }
